Squashed commit of the following:

commit d69581dbae55d16d27ee72ed5a07b90f2bd17cc5
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Wed Jun 24 14:28:10 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Add const with min supported version and use it in IgnoreIfVersionLowerThan attribute

commit 01f344fb61ebcefba9d8821fe687e04310836d17
Merge: ade35b5d3 2bf76d8b7
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Mon Jun 22 08:22:15 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Merge develop

commit ade35b5d38f5ce08888d45c3278983b879fe3796
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Thu Jun 18 16:19:28 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Add attribute for ignored tests depending on tested version,
    Add enum for Relativity versions,
    Rename RelativityVersionCompatibilityChecker into RelativityVersionChecker

commit f523e4234d3b6e1d0cbb52500614b5759d26971a
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Wed Jun 17 15:18:49 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Revert removing NotInCompatibility test category

commit 349705dfb5fda3984ccac272dfd766bd1998363f
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Wed Jun 17 10:31:05 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Turn off ChoicesTests for not compatible versions

commit a818ee24b37053b6d201538c098b5b3a5ee65d67
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Tue Jun 16 14:55:50 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Turn off tests for not compatible versions

commit 39242eb406d8ee51b6bb4fcd25eda549eb31efb2
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Tue Jun 16 14:00:58 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Add Integration category to all Integration tests - turn them on on copatibility pipeline

commit 4aac40ea96437349946679604bea0eaa59cf3835
Merge: d2e4dead5 f9ffbb03c
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Tue Jun 16 13:49:50 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Merge develop

commit d2e4dead570cd8630e84b3aa782428094166b604
Merge: cfeb9ec15 05d40f911
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Mon Jun 15 17:17:03 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Merge develop, make GetCurrentRelativityVersion public

commit cfeb9ec1582a2d5aa409bd6be6bbc6c0c9f39f03
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Mon Jun 15 13:47:48 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Remove NotInCompatibility test category,
    Turn off tests from this category only on not supported versions

commit c35b1640be7b3818cf80a3ed88531b271c34bccf
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Mon Jun 15 11:35:32 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Add message for passed tests

commit 131e7536cf3a004b42655394b63164a9ff819de2
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Mon Jun 15 10:14:54 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Add testing lanceleaf

commit 9f9d2a9bb4ad511c1209cb4350918fe39aba4e19
Author: Małgorzata Kwiatkowska <malgorzata.kwiatkowska@relativity.com>
Date:   Fri Jun 12 16:29:06 2020 +0200

    REL-439651-investigate-nightly-pipeline-tests

    Turn off tests on not compatible versions: workaround for issue in jira: 421780

+++ Source/Relativity.DataExchange.Import.NUnit.Integration/ChoicesTests.cs
@@ -7,18 +7,17 @@ namespace Relativity.DataExchange.Import.NUnit.Integration
 	using System;
 	using System.Collections.Generic;
 	using System.Linq;
-	using System.Runtime.CompilerServices;
 	using System.Threading.Tasks;
-
 	using global::NUnit.Framework;
-
 	using kCura.Relativity.DataReaderClient;
 	using Relativity.DataExchange.Import.NUnit.Integration.Dto;
 	using Relativity.DataExchange.TestFramework;
 	using Relativity.DataExchange.TestFramework.Import.JobExecutionContext;
 	using Relativity.DataExchange.TestFramework.Import.SimpleFieldsImport;
 	using Relativity.DataExchange.TestFramework.Import.SimpleFieldsImport.FieldValueSources;
+	using Relativity.DataExchange.TestFramework.NUnitExtensions;
 	using Relativity.DataExchange.TestFramework.RelativityHelpers;
+	using Relativity.DataExchange.TestFramework.RelativityVersions;
 	using Relativity.Services.Interfaces.Field.Models;
 	using Relativity.Services.Interfaces.Shared.Models;
 	using Relativity.Testing.Identification;
@@ -29,7 +28,10 @@ namespace Relativity.DataExchange.Import.NUnit.Integration
 	public class ChoicesTests : ImportJobTestBase<NativeImportExecutionContext>
 	{
 		private const int NumberOfDocumentsToImport = 10;
+		private const RelativityVersion MinSupportedVersion = RelativityVersion.Goatsbeard;
 		private readonly string multiValueDelimiter = SettingsConstants.DefaultMultiValueDelimiter.ToString();
+
+		private bool testsSkipped = false;
 
 		private int originalImportBatchSize;
 		private List<int> createdFieldsIds = new List<int>();
@@ -39,35 +41,49 @@ namespace Relativity.DataExchange.Import.NUnit.Integration
 		[OneTimeSetUp]
 		public Task OneTimeSetUpAsync()
 		{
-			this.originalImportBatchSize = AppSettings.Instance.ImportBatchSize;
-			AppSettings.Instance.ImportBatchSize = 4;
+			testsSkipped = RelativityVersionChecker.VersionIsLowerThan(this.TestParameters, MinSupportedVersion);
+			if (!testsSkipped)
+			{
+				this.originalImportBatchSize = AppSettings.Instance.ImportBatchSize;
+				AppSettings.Instance.ImportBatchSize = 4;
 
-			choicesValidator = new ChoicesValidator(this.TestParameters);
+				choicesValidator = new ChoicesValidator(this.TestParameters);
 
-			return RdoHelper.DeleteAllObjectsByTypeAsync(this.TestParameters, (int)ArtifactType.Document);
+				return RdoHelper.DeleteAllObjectsByTypeAsync(this.TestParameters, (int)ArtifactType.Document);
+			}
+
+			return Task.CompletedTask;
 		}
 
 		[OneTimeTearDown]
 		public void OneTimeTearDown()
 		{
-			AppSettings.Instance.ImportBatchSize = this.originalImportBatchSize;
+			if (!testsSkipped)
+			{
+				AppSettings.Instance.ImportBatchSize = this.originalImportBatchSize;
+			}
 		}
 
 		[TearDown]
 		public async Task TearDownAsync()
 		{
-			var deleteDocumentsTask = RdoHelper.DeleteAllObjectsByTypeAsync(this.TestParameters, (int)ArtifactType.Document);
-
-			foreach (var createdFieldId in this.createdFieldsIds)
+			if (!testsSkipped)
 			{
-				await FieldHelper.DeleteFieldAsync(this.TestParameters, createdFieldId).ConfigureAwait(false);
+				var deleteDocumentsTask = RdoHelper.DeleteAllObjectsByTypeAsync(this.TestParameters, (int)ArtifactType.Document);
+
+				foreach (var createdFieldId in this.createdFieldsIds)
+				{
+					await FieldHelper.DeleteFieldAsync(this.TestParameters, createdFieldId).ConfigureAwait(false);
+				}
+
+				await deleteDocumentsTask.ConfigureAwait(false);
+
+				this.createdFieldsIds = new List<int>();
 			}
-
-			await deleteDocumentsTask.ConfigureAwait(false);
-
-			this.createdFieldsIds = new List<int>();
 		}
 
+		[Category(TestCategories.Integration)]
+		[IgnoreIfVersionLowerThan(MinSupportedVersion)]
 		[IdentifiedTestCase("0e72e6f1-6c53-45ca-9d5d-d9ccb46f2195", new[] { "a", "b", "c", "d" }, OverlayBehavior.MergeAll)]
 		[IdentifiedTestCase("5fa9f2a6-3a8a-47c3-aeac-b558f3ad0c02", new[] { "a", "b", "c", "d" }, OverlayBehavior.ReplaceAll)]
 		[IdentifiedTestCase("dc3a8ee9-dbcc-4444-b480-4c2912f4ff1f", new[] { "a", "b", "c", "d" }, OverlayBehavior.UseRelativityDefaults)]
@@ -75,10 +91,14 @@ namespace Relativity.DataExchange.Import.NUnit.Integration
 		public Task ShouldReplaceExistingValuesForSingleChoiceFieldAsync(string[] initialUniqueChoices, OverlayBehavior overlayBehavior)
 			=> this.SingleChoiceTestAsync(overlayBehavior, initialUniqueChoices, uniqueChoices: new[] { "e", "f", "g", "h" });
 
+		[Category(TestCategories.Integration)]
+		[IgnoreIfVersionLowerThan(MinSupportedVersion)]
 		[IdentifiedTest("14c4060d-51da-4187-b41a-3de675684239")]
 		public Task ShouldUnlinkExistingValuesForSingleChoiceFieldAsync()
 			=> this.SingleChoiceTestAsync(OverlayBehavior.ReplaceAll, initialUniqueChoices: new[] { "a", "b", "c", "d" }, uniqueChoices: new[] { string.Empty });
 
+		[Category(TestCategories.Integration)]
+		[IgnoreIfVersionLowerThan(MinSupportedVersion)]
 		[IdentifiedTestCase("51c1ea5f-ffd3-43d4-b074-cd3db72c9301", new[] { "a", "b", "c", "d" }, Services.Interfaces.Field.Models.OverlayBehavior.MergeValues)]
 		[IdentifiedTestCase("2b265e75-c753-485d-bf9d-fe5e4568e752", new[] { "a", "" }, Services.Interfaces.Field.Models.OverlayBehavior.ReplaceValues)]
 		public Task ShouldReplaceExistingMultiChoicesValues(string[] initialUniqueChoices, Services.Interfaces.Field.Models.OverlayBehavior fieldOverlayBehavior) => this.MultiChoiceTestAsync(
@@ -88,6 +108,8 @@ namespace Relativity.DataExchange.Import.NUnit.Integration
 			uniqueChoices: new[] { "e", "f", "g", "h" },
 			expectedShouldMergeValues: false);
 
+		[Category(TestCategories.Integration)]
+		[IgnoreIfVersionLowerThan(MinSupportedVersion)]
 		[IdentifiedTestCase("51c1ea5f-ffd3-43d4-b074-cd3db72c9301", new[] { "a", "b", "c", "d" }, new[] { "" })]
 		[IdentifiedTestCase("a8b602ba-c305-4185-8c31-b901388a15c4", new[] { "a", "b", "c", "d" }, new[] { "a", "" })]
 		[IdentifiedTestCase("72dbd4e1-48d4-431e-89d9-5107cf096b85", new[] { "a", "" }, new[] { "e", "" })]
@@ -98,6 +120,8 @@ namespace Relativity.DataExchange.Import.NUnit.Integration
 			uniqueChoices,
 			expectedShouldMergeValues: false);
 
+		[Category(TestCategories.Integration)]
+		[IgnoreIfVersionLowerThan(MinSupportedVersion)]
 		[IdentifiedTestCase("507a2802-7003-442f-8715-a3d1539410a8", new[] { "a", "b", "c", "d" }, Services.Interfaces.Field.Models.OverlayBehavior.ReplaceValues)]
 		[IdentifiedTestCase("9c5e6380-cdf9-44c5-b730-148b22043ce9", new[] { "a", "" }, Services.Interfaces.Field.Models.OverlayBehavior.MergeValues)]
 		public Task ShouldMergeExistingMultiChoicesWithNewValues(string[] initialUniqueChoices, Services.Interfaces.Field.Models.OverlayBehavior fieldOverlayBehavior)
@@ -108,6 +132,8 @@ namespace Relativity.DataExchange.Import.NUnit.Integration
 				uniqueChoices: new[] { "e", "f", "g", "h" },
 				expectedShouldMergeValues: true);
 
+		[Category(TestCategories.Integration)]
+		[IgnoreIfVersionLowerThan(MinSupportedVersion)]
 		[IdentifiedTestCase("ee4acebb-e5bd-4d70-a7b1-8474258972e6", Services.Interfaces.Field.Models.OverlayBehavior.MergeValues, true)]
 		[IdentifiedTestCase("67aa07ba-12ea-4a48-b615-3f767114a0d1", Services.Interfaces.Field.Models.OverlayBehavior.ReplaceValues, false)]
 		public Task ShouldUseDefaultMergeBehaviorForMultiChoices(Services.Interfaces.Field.Models.OverlayBehavior fieldOverlayBehavior, bool expectedShouldMergeValues)
@@ -118,6 +144,8 @@ namespace Relativity.DataExchange.Import.NUnit.Integration
 				uniqueChoices: new[] { "e", "f", "g", "h" },
 				expectedShouldMergeValues);
 
+		[Category(TestCategories.Integration)]
+		[IgnoreIfVersionLowerThan(MinSupportedVersion)]
 		[IdentifiedTestCase("d3809802-e166-49a2-a30b-30b0cb53eef2", NumberOfDocumentsToImport)]
 		[IdentifiedTestCase("9c5952f2-5098-4ad4-876c-79109a369ae2", NumberOfDocumentsToImport / 2)]
 		public async Task ShouldImportSingleAndMultiChoiceInOneJob(int numberOfDocumentToImportInImportUnderTest)
